+++ src/Providers/CommandServiceProvider.php
@@ -5,7 +5,6 @@ namespace Tapestry\Providers;
 use Tapestry\Tapestry;
 use Tapestry\Console\Application;
 use Tapestry\Console\Commands\InitCommand;
-use Tapestry\Console\Commands\MakeCommand;
 use Tapestry\Console\Commands\BuildCommand;
 use Tapestry\Console\Commands\SelfUpdateCommand;
 use League\Container\ServiceProvider\AbstractServiceProvider;
@@ -56,6 +55,7 @@ class CommandServiceProvider extends AbstractServiceProvider
                     \Symfony\Component\Filesystem\Filesystem::class,
                     \Symfony\Component\Finder\Finder::class,
                 ]);
+
             array_push($commands, $container->get(SelfUpdateCommand::class));
         }