+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Tapestry currently operates a one month release cycle between releases, see the 
 
 ### Fixed
 - #272 Disable self-update command if not executed within a phar
+- #148 Frontmatter is now parsed in files with an empty body.
 
 ### Added
 - #157 Added lock file support so Tapestry doesn't run concurrently